pyproject.toml
[tool.poetry]
name = "video-generator"
version = "0.1.0"
description = "Combines images and audio to create a video"
authors = ["1blckhrt <williams.1691@wright.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.13.1"
pillow = "^11.0.0"
moviepy = "^2.2.1"


[tool.poetry.group.dev.dependencies]
poethepoet = "^0.31.1"
pre-commit = "^4.0.1"
commitlint = "^1.3.0"
black = "^24.10.0"
flake8 = "^7.1.1"
pyinstaller = "^6.11.1"

[tool.poe.tasks]
format = "black src/*.py"
lint = "flake8 src/*.py"
dev = "py src/main.py"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
